[Refactor][Ops] delete LerpTensorFwdOp fake-impl and supporting artifacts (#1250)
Closes #1237

## Summary

- Delete `LerpTensorFwdOp` fake-impl class,
`_register_lerp_tensor_custom_op`, and `__all__` entry from
`tileops/ops/elementwise.py`.
- Delete `tests/ops/test_elementwise_multi_input_alignment.py` and
`benchmarks/ops/bench_elementwise_multi_input.py` (entire fake-impl
scaffolding).
- Remove `LerpTensorFwdOp` import + `test_lerp_tensor_compile` from
`tests/ops/test_elementwise_compile.py`.
- Move WhereFwdOp fp8 reject + positive-dtype tests into
`tests/ops/test_elementwise_independent_fp8.py`.
- Manifest entry stays `status: spec-only` (byte-identical, sha256
unchanged); LerpTensorFwdOp now referenced only by manifest YAML.

Follow-up issue #1249 tracks the real `LerpTensorFwdOp` implementation.

## Test plan

- [x] AC-1: pytest tests/ops/ green (271/271 elementwise smoke tests
pass at HEAD; 37 unrelated failures on shared host trace to
`/tmp/tvm-debug-mode-tempdirs` permission, not this diff — same env
issue as #1192/#1196).
- [x] AC-2: `LerpTensorFwdOp` class + `_register_lerp_tensor_custom_op`
+ `__all__` entry removed; `grep -nE
'LerpTensorFwdOp|_register_lerp_tensor_custom_op'
tileops/ops/elementwise.py` returns no matches.
- [x] AC-3: `benchmarks/ops/bench_elementwise_multi_input.py` deleted.
- [x] AC-4: `tests/ops/test_elementwise_multi_input_alignment.py`
deleted; WhereFwdOp fp8 reject + positive dtype tests relocated to
`tests/ops/test_elementwise_independent_fp8.py`.
- [x] AC-5: `LerpTensorFwdOp` import + `test_lerp_tensor_compile`
removed from `tests/ops/test_elementwise_compile.py`.
- [x] AC-6: `tileops/manifest/elementwise_multi_input.yaml`
byte-identical (sha256 `cbb1da67…1d5e6` before == after).
- [x] AC-7: WhereFwdOp fp8 rejection preserved verbatim
(`_SUPPORTED_DTYPES` tuple + `__init__` gate); behavior covered in new
test location.
- [x] AC-8: Follow-up issue #1249 filed for real `LerpTensorFwdOp` impl.
- [ ] AC-9: `grep -rn 'LerpTensorFwdOp' tileops/ tests/ benchmarks/
--exclude-dir=manifest` returns one match —
`tileops/perf/formulas.py:447` docstring "Roofline for LerpTensorFwdOp
(Tensor-weight torch.lerp)" documenting the manifest's `roofline.source`
symbol. Resolved in favor of AC-10 below; perf docstring cleanup
deferred to land alongside #1249.
- [x] AC-10: Diff confined to the five files listed in the goal; no
manifest, kernels, docs, .claude, or perf changes.

## Structural Readiness

All checks passed.

## Test node delta

```
File                                             Base    HEAD    Delta
----------------------------------------------------------------------
tests/ops/test_elementwise_compile.py               0      73      +73
tests/ops/test_elementwise_independent_fp8.py      59      64       +5
----------------------------------------------------------------------
TOTAL                                              59     137      +78
```

**Justification:** The +73 on `test_elementwise_compile.py` is a
measurement artifact — `test_node_delta.py` reported "pytest collection
failed" on the base ref (existing pre-PR state), so its base count of 0
is wrong. The real net change is: `test_elementwise_compile.py` loses
one node (removed `test_lerp_tensor_compile`);
`test_elementwise_independent_fp8.py` gains 5 nodes (two WhereFwdOp
behavior tests moved in from the deleted
`test_elementwise_multi_input_alignment.py` file, parametrized over fp8
dtypes). Overall the PR is net-negative on test node count (one entire
test file deleted; 5 nodes relocated).
