+++ tileops/ops/elementwise.py
@@ -283,44 +283,6 @@ def _register_where_custom_op(op_cls):
     op_cls._wrapped = _wrapped
 
 
-def _register_lerp_tensor_custom_op(op_cls):
-    """Register a Tensor-weight lerp op (input, end, weight -> out).
-
-    The fake function computes the broadcast output shape from ``input`` /
-    ``end`` / ``weight`` so that ``torch.compile(fullgraph=True)`` works
-    for both same-shape and broadcasting inputs. The op is registered
-    as ``top::elementwise_{op_name}`` (e.g. ``top::elementwise_lerp_tensor``
-    when ``op_cls._op_name == 'lerp_tensor'``), giving it a distinct name
-    from the scalar ``LerpFwdOp`` (registered as ``top::elementwise_binary_lerp``
-    via the binary factory) so the two overloads cannot collide.
-    """
-    op_name = op_cls._op_name
-
-    @torch.library.custom_op(
-        f"top::elementwise_{op_name}", mutates_args=(),
-    )
-    def _wrapped(
-        input: torch.Tensor,  # noqa: A002
-        end: torch.Tensor,
-        weight: torch.Tensor,
-        instance_key: int,
-    ) -> torch.Tensor:
-        instance = _OP_REGISTRY[instance_key]
-        return instance._eager_forward(input, end, weight)
-
-    @_wrapped.register_fake
-    def _(
-        input: torch.Tensor,  # noqa: A002
-        end: torch.Tensor,
-        weight: torch.Tensor,
-        instance_key: int,
-    ) -> torch.Tensor:
-        out_shape = torch.broadcast_shapes(input.shape, end.shape, weight.shape)
-        return input.new_empty(out_shape)
-
-    op_cls._wrapped = _wrapped
-
-
 def _register_masked_fill_custom_op(op_cls):
     """Register a masked-fill-style op (x, mask -> y) for torch.compile.
 
@@ -622,7 +584,6 @@ __all__ = [
     "SoftplusFwdOp",
     "PreluFwdOp",
     "WhereFwdOp",
-    "LerpTensorFwdOp",
     "ClampFwdOp",
     "ClampScalarFwdOp",
     "ClampMinFwdOp",
@@ -2401,98 +2362,6 @@ class WhereFwdOp(Op):
         return self._eager_forward(condition, input, other)
 
 
-class LerpTensorFwdOp(Op):
-    """Tensor-weight lerp: out = input + weight * (end - input).
-
-    Conforms to the Tensor-weight overload of ``torch.lerp`` —
-    ``torch.lerp(input, end, weight: Tensor)`` where ``weight`` is a
-    Tensor that broadcasts together with ``input`` and ``end`` to the
-    output shape (vs ``LerpFwdOp``'s scalar ``weight`` baked at
-    construction time). The kernel-level implementation is not yet
-    available; the op layer dispatches to ``torch.lerp`` so the manifest
-    L1 contract (signature + broadcast semantics) is honoured at the
-    op boundary while the manifest entry remains ``status: spec-only``.
-
-    Args:
-        input: Shape of the start tensor.
-        end: Shape of the end tensor.
-        weight: Shape of the per-element weight tensor.
-        dtype: Torch dtype for all three operands.
-    """
-
-    _op_name = "lerp_tensor"
-    _wrapped = None
-    # Manifest declares all three operands as ``float16 | bfloat16 | float32``.
-    _SUPPORTED_DTYPES = (torch.float16, torch.bfloat16, torch.float32)
-
-    def __init__(
-        self,
-        input: tuple,  # noqa: A002 — manifest-aligned PyTorch param name
-        end: tuple,
-        weight: tuple,
-        dtype: torch.dtype,
-    ):
-        if dtype not in self._SUPPORTED_DTYPES:
-            names = ", ".join(str(dt) for dt in self._SUPPORTED_DTYPES)
-            raise ValueError(
-                f"LerpTensorFwdOp does not support dtype {dtype}. "
-                f"Supported: [{names}]"
-            )
-        self.input_shape = tuple(input)
-        self.end_shape = tuple(end)
-        self.weight_shape = tuple(weight)
-        self.dtype = dtype
-        self.out_shape = tuple(
-            torch.broadcast_shapes(
-                self.input_shape, self.end_shape, self.weight_shape,
-            )
-        )
-        self.N_total = prod(self.out_shape) if self.out_shape else 1
-        self._instance_key = id(self)
-        _OP_REGISTRY[self._instance_key] = self
-
-    @property
-    def default_kernel_map(self) -> Dict[str, Kernel]:
-        # Manifest entry has no ``kernel_map`` (spec-only). Returning an
-        # empty dict honours the abstract-method contract; callers must
-        # not invoke ``dispatch_kernel`` on this op until a kernel ships.
-        return {}
-
-    def _eager_forward(
-        self,
-        input: torch.Tensor,  # noqa: A002 — manifest-aligned PyTorch param name
-        end: torch.Tensor,
-        weight: torch.Tensor,
-    ) -> torch.Tensor:
-        return torch.lerp(input, end, weight)
-
-    def forward(
-        self,
-        input: torch.Tensor,  # noqa: A002 — manifest-aligned PyTorch param name
-        end: torch.Tensor,
-        weight: torch.Tensor,
-    ) -> torch.Tensor:
-        if not (input.is_cuda and end.is_cuda and weight.is_cuda):
-            raise ValueError("Inputs must be CUDA tensors")
-        for name, t, expected in [
-            ("input", input, self.input_shape),
-            ("end", end, self.end_shape),
-            ("weight", weight, self.weight_shape),
-        ]:
-            if t.dtype != self.dtype:
-                raise ValueError(
-                    f"Expected {name}.dtype {self.dtype}, got {t.dtype}"
-                )
-            if tuple(t.shape) != expected:
-                raise ValueError(
-                    f"Expected {name}.shape {expected}, got {tuple(t.shape)}"
-                )
-        wrapped = type(self)._wrapped
-        if wrapped is not None:
-            return wrapped(input, end, weight, self._instance_key)
-        return self._eager_forward(input, end, weight)
-
-
 class _ClampTensorBase(Op):
     """Shared infrastructure for Tensor-bound clamp variants (broadcasting)."""
 
@@ -3288,13 +3157,6 @@ _register_clamp_max_custom_op(ClampMaxFwdOp)
 # with the scalar variant's ``top::elementwise_masked_fill``.
 _register_masked_fill_custom_op(MaskedFillScalarFwdOp)
 _register_masked_fill_tensor_value_custom_op(MaskedFillFwdOp)
-
-# --- Tensor-weight lerp (1 op: input, end, weight -> out) ---
-# Registered under ``top::elementwise_lerp_tensor`` to avoid colliding with
-# the scalar ``LerpFwdOp``'s ``top::elementwise_binary_lerp`` namespace. The fake
-# function is broadcast-aware so torch.compile(fullgraph=True) traces
-# correctly for both same-shape and broadcasting inputs.
-_register_lerp_tensor_custom_op(LerpTensorFwdOp)
 
 # --- Where op (1 op: cond, x, y -> out) ---
 # The fake function is broadcast-aware so torch.compile(fullgraph=True)
